+++ src/main/java/com/leonyk2/datagen/ModRecipeProvider.java
@@ -2,19 +2,19 @@ package com.leonyk2.datagen;
 
 import com.leonyk2.mcmod.McMod;
 import com.leonyk2.mcmod.item.ModItems;
+import com.simibubi.create.content.kinetics.press.PressingRecipe;
+import com.simibubi.create.content.processing.sequenced.SequencedAssemblyRecipeBuilder;
 import net.minecraft.data.PackOutput;
 import net.minecraft.data.recipes.*;
-import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.item.Items;
-import net.minecraft.world.item.crafting.AbstractCookingRecipe;
-import net.minecraft.world.item.crafting.Ingredient;
-import net.minecraft.world.item.crafting.RecipeSerializer;
-import net.minecraft.world.level.ItemLike;
 import net.minecraftforge.common.crafting.conditions.IConditionBuilder;
 import org.jetbrains.annotations.NotNull;
 
 import java.util.List;
 import java.util.function.Consumer;
+import java.util.function.UnaryOperator;
+
+import static net.minecraft.resources.ResourceLocation.fromNamespaceAndPath;
 
 public class ModRecipeProvider extends RecipeProvider implements IConditionBuilder {
     public ModRecipeProvider(PackOutput pOutput) { super(pOutput); }
@@ -59,28 +59,35 @@ public class ModRecipeProvider extends RecipeProvider implements IConditionBuild
                 .pattern("CCC")
                 .define('C', Items.CHARCOAL)
                 .unlockedBy(getHasName(Items.CHARCOAL), has(Items.CHARCOAL))
-                .save(pWriter, ResourceLocation.fromNamespaceAndPath(McMod.MOD_ID, "coal_block"));
+                .save(pWriter, fromNamespaceAndPath(McMod.MOD_ID, "coal_block"));
         ShapelessRecipeBuilder.shapeless(RecipeCategory.MISC, Items.QUARTZ, 4)
                 .requires(Items.QUARTZ_BLOCK)
                 .unlockedBy(getHasName(Items.QUARTZ_BLOCK), has(Items.QUARTZ_BLOCK))
-                .save(pWriter, ResourceLocation.fromNamespaceAndPath(McMod.MOD_ID, "quartz"));
-    }
+                .save(pWriter, fromNamespaceAndPath(McMod.MOD_ID, "quartz"));
 
 
-    protected static void oreSmelting(@NotNull Consumer<FinishedRecipe> pFinishedRecipeConsumer, List<ItemLike> pIngredients, @NotNull RecipeCategory pCategory, @NotNull ItemLike pResult, float pExperience, int pCookingTIme, @NotNull String pGroup) {
-        oreCooking(pFinishedRecipeConsumer, RecipeSerializer.SMELTING_RECIPE, pIngredients, pCategory, pResult, pExperience, pCookingTIme, pGroup, "_from_smelting");
-    }
+        SequencedAssemblyRecipeBuilder builder =
+                new SequencedAssemblyRecipeBuilder(fromNamespaceAndPath(McMod.MOD_ID, "diamond_from_pressing_sequence_coal"));
+        builder.require(Items.COAL);
+        builder.transitionTo(ModItems.UNCOMPLETE_DIAMOND.get());
+        builder.addStep(PressingRecipe::new, UnaryOperator.identity());
+        builder.addStep(PressingRecipe::new, UnaryOperator.identity());
+        builder.addStep(PressingRecipe::new, UnaryOperator.identity());
+        builder.addStep(PressingRecipe::new, UnaryOperator.identity());
+        builder.addStep(PressingRecipe::new, UnaryOperator.identity());
+        builder.addOutput(Items.DIAMOND, 1.0f);
+        builder.build(pWriter);
 
-    protected static void oreBlasting(@NotNull Consumer<FinishedRecipe> pFinishedRecipeConsumer, List<ItemLike> pIngredients, @NotNull RecipeCategory pCategory, @NotNull ItemLike pResult, float pExperience, int pCookingTime, @NotNull String pGroup) {
-        oreCooking(pFinishedRecipeConsumer, RecipeSerializer.BLASTING_RECIPE, pIngredients, pCategory, pResult, pExperience, pCookingTime, pGroup, "_from_blasting");
-    }
-
-    protected static void oreCooking(@NotNull Consumer<FinishedRecipe> pFinishedRecipeConsumer, @NotNull RecipeSerializer<? extends AbstractCookingRecipe> pCookingSerializer, List<ItemLike> pIngredients, @NotNull RecipeCategory pCategory, @NotNull ItemLike pResult, float pExperience, int pCookingTime, @NotNull String pGroup, String pRecipeName) {
-        for(ItemLike itemlike : pIngredients) {
-            SimpleCookingRecipeBuilder.generic(Ingredient.of(itemlike), pCategory, pResult,
-                            pExperience, pCookingTime, pCookingSerializer)
-                    .group(pGroup).unlockedBy(getHasName(itemlike), has(itemlike))
-                    .save(pFinishedRecipeConsumer,  McMod.MOD_ID + ":" + getItemName(pResult) + pRecipeName + "_" + getItemName(itemlike));
-        }
+        SequencedAssemblyRecipeBuilder builder2 =
+                new SequencedAssemblyRecipeBuilder(fromNamespaceAndPath(McMod.MOD_ID, "diamond_block_from_pressing_sequence_coal"));
+        builder2.require(Items.COAL_BLOCK);
+        builder2.transitionTo(ModItems.UNCOMPLETE_DIAMOND.get());
+        builder2.addStep(PressingRecipe::new, UnaryOperator.identity());
+        builder2.addStep(PressingRecipe::new, UnaryOperator.identity());
+        builder2.addStep(PressingRecipe::new, UnaryOperator.identity());
+        builder2.addStep(PressingRecipe::new, UnaryOperator.identity());
+        builder2.addStep(PressingRecipe::new, UnaryOperator.identity());
+        builder2.addOutput(Items.DIAMOND_BLOCK, 1.0f);
+        builder2.build(pWriter);
     }
 }